+++ certlogic/certlogic-js/src/misc/index.ts
@@ -1,2 +1,2 @@
-export { desugar } from "./desugar"
+export { desugar, or_ } from "./desugar"
 export { renderAsCompactText } from "./textual-notation"